test(e2e): cover impersonation read-only on CSE static path (#3765)
